Fix wrong name placement for some classifiers
For classifiers like the base or the 'frame' class,
the 'name' text element was placed overlapping
the right edge of the surrounding shape.
This was due to the style attribute of text elements being
overwritten is some cases.

If a custom strokeColor is set, it is now correctly appended
to the style attribute of text elements instead of overwriting it.